Saving player-, server-, world-Data. not chunks though

+++ src/main/java/server/Server.java
@@ -14,7 +14,7 @@ import static utils.Constants.*;
 
 public final class Server {
 
-    public Server(int currentGameTick) {
+    public Server(long currentGameTick) {
         this.currentGameTick = currentGameTick;
     }
